add admin function and banner message
OK, if I want to display a message to users, this is an easy and quick way to do it.

now I don't need to reploy or anything like that, I can just come to this /admin route, type out a message, the message gets saved to the message.txt file. if the file is empty, the div containing the banner message is erased via javascript on the front end. if there is something in the message.txt file, it will show at the top of the page