passing more info to the front end
ok. so we want to be able to sort things on the front end. we're passing in in this info. it's based solely on the title of the podcast, which gets its name from the datetime - IE the time the recording begins.

this can be accessed via jinga. we're adding this info as a class, then we can sort it later with JS.

it's late, i'm tired, hope that makes sense.