move more JS to file, remove uneeded JS
moving some more JS functions over to the main.js file in the static directory.

-big block of deleted JS
we had that one JS function on the main page, which would check all of the input checkboxes when the page loaded. we needed this when we were using bare, native html checkboxes. now that we're using these cool bootstrap label button things, the indication to the user of whether or not the checkbox is checked is when the color changes. so it doesn't actually matter if the checkboxes are 'checked' when the page loads, since their functionality is really more about what happens when they are clicked, not whether they are checked or not.

also moving the script tag for the main.js file up to the top of the page. apparently it matters that this is located. if an element loads before the JS that maipulates it is loaded, it will load the element empty.