fix banner
previously, we were always rendering the banner, then, with JS, if the banner message was empty, we'd remove it from the DOM. this was fine, but sometimes the element would load briefly and be visible on the screen before JS removed it.

now, we're using jinja to decide whether to render the banner element at all.