css for toggle  buttons
so this is a lot of audio files, and they can take a few moments to render. if you start clicking the toggle buttons right away, sometimes they are not all fully loaded yet, and it messes up the toggle switches. this adds a nice fade in for the toggle buttons, and also ensures they take a little longer to load.

TODO, don't load the toggle buttons until you're sure the audio (and audio elements) are all fully loaded with proper class names and such.