hide stop button when no audio playing
previously, there was a bug with the stop button. it would display when user clicked play. it would go away when user clicked the stop button. but if user clicked pause button on an individual player, the stop button would just sit there. you cannot use an "pause" event to hide it, because clicking play on any audio element is pausing all other elements.

anyways, now we finally have a way to hide it. when user clicks pause (onpause()), it calls this new function. the function checks whether all the audio elements on the page are paused. if ALL of them are paused, the hide button is hidden. if all are not paused, nothing happens.