turning preload back on
when this is off, the player does not show the length. i wrote a function to show the length before hand, but then when you click to play the file, the length comes up in the audio player, so now the length is listed twice. so it's either that, or have all the audio files loaded when the page loads. neither option is appealing...so just leaving as it is for now.