remove this fixture
testing if I can remove this function. The EV is only accessed when the real class method is called. since we are mocking the method, the real method should never be called and thus the system should never reach out to grab the EV. RIGHT??? We shall see. Pushing this to GH Actions to see if the tests pass...