use dictionary instead of list for episodes
previously we were using a list of lists and sending that to Jinja. Then to access things in Jinja we were using the index value of each item. When looking at it this way, we have no way of knowing what's at each index. With the dictionary, we can access each item with its key name, for easier readability.

so it's a list of dictionaries. loop through them and grab the key.