fix audio player border issue
when using a border radius, the corners of the background are still shown. when you use overflow:hidden, the background goes away, BUT it cuts off the progress bar which sits above the border.

need to come back to this to get rounded edges working with the progress bar showing